Sync with the latest code
This patch will override some previous patches.

Socket activation code for systemd is not good for this package.
It should not be enabled from this library.

User (who will develop S/W using this package) should have chance,
To use the socket activation feature or not.

It should be implemented from more high layer.

Change-Id: Ic3df7762f0a86f5984d7e9a0238baaaed07c535c
Signed-off-by: Sung-jae Park <nicesj@nicesj.com>