Sync with the latest product code of phone profile
Disable O_NONBLOCK option for creating a "pipe".
It makes unpredictable busy waiting problem when it fails to write data to pipe.
In some cases, the consumer thread could not empty the pipe,
the producer thread failed to write with EAGAIN error if the pipe created with O_NONBLOCK.
then the producer thread fall into the busy waiting status.

This patch will keep the producer from busy waiting problem.

Change-Id: I34d1433b4c380deb2bc8f03d476943652de6c23f