BUFSIZ is replaced with payload_size
At least now, while I make a product using this code.
There is no such problem what I read from history in Tizen 3.0.
If the crash occurred when allocating heap more large size than BUFSIZ,
It mean somewhere else is broken.

It could be caller of this function.

The problem should be invetigated again.

If you really think that the problem is caused by this.
Please let me know it. contact to me.
We need discuss it.

[model] Redwood,Kiran,B3(Wearable)
[binary_type] AP
[customer] Docomo/Orange/ATT/Open
[issue#] N/A
[problem]
[cause]
[solution]
[team] HomeTF
[request]
[horizontal_expansion]

Change-Id: I3b3c46b7c023a5805a4bb63081049e9a534e9ecc